requirements.txt
dnspython>=2.3.0
dnslib>=0.9.20
Flask>=2.3.0
pytest>=7.0.0
cachetools>=5.0.0
